fix: fixed the problem that when typing, the order of the messages reverses
