+++ src/components/Chat/index.tsx
@@ -1,13 +1,11 @@
-import React, { useState } from "react"
+import React from "react"
 import { NavigateFunction, useNavigate } from "react-router-dom"
 import {
     DocumentData,
-    addDoc,
     collection,
     limit,
     orderBy,
-    query,
-    serverTimestamp
+    query
 } from "firebase/firestore"
 import { useCollectionData } from "react-firebase-hooks/firestore"
 import { IDOptions, InitialValueOptions } from "react-firebase-hooks/firestore/dist/firestore/types"
@@ -15,14 +13,11 @@ import { IDOptions, InitialValueOptions } from "react-firebase-hooks/firestore/d
 import {
     Avatar,
     Content,
-    InputContent,
-    SendBtn,
-    SendBtnIcon,
     SettingsIcon
 } from "./styles"
 
-import Input from "../Input"
 import Header from "../Header"
+import ChatInput from "./ChatInput"
 import ScreenTitle from "../ScreenTitle"
 import LeftHeader from "../Header/LeftHeader"
 import RightHeader from "../Header/RightHeader"
@@ -34,25 +29,7 @@ import { auth, databaseApp } from "../../services/firebaseConfig"
 const Chat: React.FC = () => {
     const navigate: NavigateFunction = useNavigate()
 
-    const [text, setText] = useState("")
-    
     const messageRef = collection(databaseApp, "messages")
-
-    const sendMessage = async (text: string) => {
-            await addDoc(messageRef, {
-                text: text,
-                uid: auth.currentUser?.uid,
-                username: auth.currentUser?.displayName,
-                createdAt: serverTimestamp()
-            })
-        }
-
-    const send = (text: string) => {
-        if (text.trim().length > 0) {
-            sendMessage(text)
-            setText("")
-        }
-    }
         
     const queryMessages = query(
         messageRef,
@@ -79,27 +56,7 @@ const Chat: React.FC = () => {
                 </RightHeader>
             </Header>
             <Messages messages={messages?.reverse()!}/>
-            <InputContent>
-                <Input
-                    placeholder="Digite sua mensagem..."
-                    value={text}
-                    onChange={
-                        (event: React.ChangeEvent<HTMLInputElement>) => setText(event.target.value)
-                    }
-                    onKeyDown={
-                        (event: React.KeyboardEvent<HTMLInputElement>) => {
-                            if (event.key === "Enter" && !event.shiftKey) send(text)
-                        }
-                    }
-                />
-                <SendBtn>
-                    <SendBtnIcon
-                        onClick={
-                            () => send(text)
-                        }
-                    />
-                    </SendBtn>
-            </InputContent>
+            <ChatInput />
         </Content>
     )
 }
